should ACTUALLY work now. Basically added attack roll and defense roll capabilites. Next is attack styles and enum stuff prob. learned how to use HashMap which is nice. got an rc level and stuff (25 levels to max now :D)

+++ combat/classes/Armor.java
@@ -8,8 +8,8 @@ public abstract class Armor {
     protected int crush;
     protected int range;
     protected int mage;
-    protected double meleestr;
-    protected double rangestr;
+    protected int meleestr;
+    protected int rangestr;
     protected double magestr;
     protected int stabdef;
     protected int slashdef;
@@ -43,7 +43,7 @@ public abstract class Armor {
         a.mage = mage;
     }
 
-    protected void setStrBonus(Armor a, double meleestr, double rangestr, double magestr)
+    protected void setStrBonus(Armor a, int meleestr, int rangestr, double magestr)
     {
         a.meleestr = meleestr;
         a.rangestr = rangestr;
